framework: compute patch in a single step
As we noticed in our operators we often do not really need update, or
want to postpone implementation. E.g. we do not support rolling
certificates, but we want to add that feature later. This require being
able to specify that update part should not be invoked by the framework.

Introduction of Patch struct is a solution to that problem. Create,
Delete and Update functions are called by the framework only when patch
has corresponding parts set.

The Patch is computed in similar way to previous GetUpdateState
function, which is now NewUpdatePatch. NewUpdatePatch function creates
a Patch object when CRO is created or updated. There is also
NewDeletePatch function which is called when CRO is deleted. Both
NewUpdatePatch and NewDeletePatch create Patch struct which in turn is
processed by Create, Delete and Update functions of the Resource.

ProcessAdd is redirected to ProcessUpdate.

ProcessUpdate has following flow:

    current = resource.GetCurrentState(cro)
    desired = resource.GetDesiredState(cro)
    patch = resource.NewUpdatePatch(cro, current, desired)
    patch.hasCreate && resource.Create(cro, patch.getCreate)
    patch.hasDelete && resource.Delete(cro, patch.getDelete)
    patch.hasUpdate && resource.Delete(cro, patch.getUpdate)

ProcessUpdate has following flow:

    current = resource.GetCurrentState(cro)
    patch = resource.NewUpdatePatch(cro, current)
    patch.hasCreate && resource.Create(cro, patch.getCreate)
    patch.hasDelete && resource.Delete(cro, patch.getDelete)
    patch.hasUpdate && resource.Delete(cro, patch.getUpdate)