framework: introduce Patch struct (#92)
As we noticed in our operators we often do not really need update, or
want to postpone implementation. E.g. we do not support rolling
certificates, but we want to add that feature later. This require being
able to specify that update part should not be invoked by the framework.

Introduction of Patch struct is a solution to that problem. ApplyCreateChange,
ApplyDeleteChange and ApplyUpdateChange functions are called by the framework
only when patch has corresponding parts set.

The Patch is computed in similar way to previous GetUpdateState function, which
is now NewUpdatePatch. NewUpdatePatch function creates a Patch object when
custom object is created or updated. There is also NewDeletePatch function
which is called when custom ojbject is deleted. Both NewUpdatePatch and
NewDeletePatch create Patch struct which in turn is processed by
ApplyCreateChange, ApplyDeleteChange and ApplyUpdateChange functions of the
Resource.

ProcessCreate is redirected to ProcessUpdate.

ProcessUpdate has following flow:

    current = resource.GetCurrentState(cro)
    desired = resource.GetDesiredState(cro)
    patch = resource.NewUpdatePatch(cro, current, desired)
    patch.hasCreate && resource.ApplyCreateChange(cro, patch.getCreate)
    patch.hasDelete && resource.ApplyDeleteChange(cro, patch.getDelete)
    patch.hasUpdate && resource.ApplyUpdateChange(cro, patch.getUpdate)

ProcessUpdate has following flow:

    current = resource.GetCurrentState(cro)
    patch = resource.NewUpdatePatch(cro, current)
    patch.hasCreate && resource.ApplyCreateChange(cro, patch.getCreate)
    patch.hasDelete && resource.ApplyDeleteChange(cro, patch.getDelete)
    patch.hasUpdate && resource.ApplyUpdateChange(cro, patch.getUpdate)