framework: allow desired state for deleted object
After tyring to apply latest operatorkit changes I discovered it would
be very useful if the desired state was created with the knowledge if
the custom object was created or deleted. This change adds additional
argument `deleted` to GetDesiredState and GetCurrentState (for
consistency). This argument indicates whether the custom object was
deleted or updated.

Examples:

cert-operator: to compute desired state we need to know if the custom
object was deleted to decide if we desire to have certificate created or
deleted.

endpoint-operator: to compute desired state we need to know if the
custom object was deleted to decide if we desire to have IPs added or
removed from endpoint subnet.