Add initial version of troubleshooting guide
Purpose of this guide is to document common mistakes that are easy to
make accidentally but hard to debug due to lack of informative error
messages.

Pitfalls listed here are stumbled on once in a while and often it might
take considerable time for multiple developer to figure out what's wrong
when error message is not perfectly clear. Therefore having a central
place to look for common mistakes as a first aid can be useful and save
time.