Don't construct new context when updating existing one
Construction of new context when updating existing one creates a memory leak
because internal implementation of reflect.WithValue() uses reflect.valueCtx
that is essentially a linked list. Therefore construction of new context on
value setting effectively adds a new element to this list and therefore creates
a memory leak when called e.g. in loop.

With this change the new context is only created when existing context doesn't
contain loggermeta context value.